refactor: getMentoringDetail api 공통으로 이동 #595
- 상세페이지와 예약 페이지에서 공통적으로 사용되어서 common으로 이동
