--- frontend/src/pages/detail/types/CertificatesResponse.ts
@@ -1,6 +1,0 @@
-export interface CertificateResponse {
-  certificateId: string;
-  title: string;
-  type: string;
-  imageUrl: string;
-}
